+++ server/src/domains/post/dto/PostDto.ts
@@ -19,45 +19,3 @@ export class PostDto {
   @IsNumber()
   hit: number;
 }
-
-export class PostContentDto {
-  @IsNumber()
-  groupId: number;
-  @IsString()
-  @MinLength(1)
-  @MaxLength(100)
-  title: string;
-  @IsString()
-  @MinLength(1)
-  @MaxLength(1023)
-  content: string;
-  @IsEnum(PostType)
-  type: PostType;
-
-  toEntity(userId: number): Post {
-    const post: Post = new Post();
-
-    post.userId = userId;
-    post.groupId = this.groupId;
-    post.title = this.title;
-    post.content = this.content;
-    post.type = this.type;
-    post.createdAt = new Date();
-    post.hit = 0;
-
-    return post;
-  }
-}
-
-export class PostUpdateDto {
-  @IsNumber()
-  id: number;
-  @IsNumber()
-  groupId: number;
-  @IsString()
-  title: string;
-  @IsString()
-  content: string;
-  @IsEnum(PostType)
-  type: PostType;
-}
